Fix  withdrawals

+++ data/src/main/java/com/a2a/data/model/card/creditcard/lasttransaction/CardLastTransactionResponse.kt
@@ -135,7 +135,7 @@ data class CardLastTransactionResponse(
                 @SerializedName("TransInAccCurr")
                 var transInAccCurr: String = "",
                 @SerializedName("Withdrawals")
-                var withdrawals: Int = 0
+                var withdrawals: Double = 0.0
             ) : Parcelable
         }
